added tests specific to redirects
